new class, private method
